Prepare items before calling layout in carousel extension
The items need to be prepared before calling layout, if they are not
prepared the collection view layout will show a warning that the item
is to big for the layout. This is because it is using the old size and
not the new one based of the new window size. This happens when you
resize the window that the component belongs too.