Apply animation after guard
Move the applyAnimation invocation so that the collection view never
ends up using the wrong animation. If the collection view were to stop
at the guard demeaning that there is nothing to update, the collection
view layout would stil store the animation which is not what we want.
Moving applying the animation fixes this issue.