Fix scrolling being disabled when starting with an empty collection.
If a horizontal component was initialized without items then scrolling
would never be enabled as it was set in setup instead of `layout`.
However, now it will check if scrolling should be enabled inside
`afterUpdate`. This should handle the cases that we need to cover for
it to always be enabled when applicable.