Refactor SpotsScrollView and SpotsContentView
The KVO implementation for observing views has been refactored to
resemble the implementation on iOS and tvOS.

`subviewsInLayoutOrder` has been moved to `SpotsContentView`. The main
reason behind this is that `NSView` does not support `func
insertSubview(UIView, at: Int)`. So when replacing `Component`, they
can end up in the wrong location. By moving the implementation over to
`SpotsContentView` we ensure that the layout order stays intact between
mutations.