Implement ComponentState
Implements `Component.state` that can be monitored by implementing
`ComponentStateDelegate`.

The `ComponentStateDelegate` requires one method:

`handleState(for component: Component, from: ComponentState, to:
ComponentState)`

It gives you the `Component` that was modified and the two state that
the `Component` switched between. The `ComponentStateDelegate` will not
be called if the `Component` changes to the same state as it had before.

Using `ComponentState` helps when handling reloading and empty states.
It is also useful for caching. You can easily trigger the `cache`
method on the `Component` when the state goes from `updating` to
`ready`.

`ComponentState` has 5 states:

initializing, setup, ready, updating, deinitializing

There might be more or less in the future.