Implement proper Interaction.mouseClick behavior for NSCollectionView
When `Interaction.mouseClick` is set to `double`, the delegate
invocation will be ignore and instead handled in `GridWrapper`.

`NSCollectionViewItem`'s have access to the `NSCollectionView` it
belongs to. It will resolve the `Component` from the `NSCollectionView`
delegate and trigger the `delegate` method `component(_ component:
Component, itemSelected: Item)` for all currently selected items in the
collection view.