Add workaround for contentInsets when using tabs
macOS 10.12 and later has support for tabs. As far as I know, there are
now good ways of checking if the window is using tabs or not and when
it does have tabs the `contentInset.top` is different. To make the
controller adapt to the new content insets we modify the width -1 and
then +1 to trigger a redrawing of the scrollview auto resizing mask so
that the new content inset is used. If we don't do this, the view will
be cut off at the top until the window is resized. This usually only
happens for the first window in the tab collection.