Refactor updating the data source using performUpdate method
Performing soft updates didn't work properly for collection views as
that update was not wrapped in a batch update. To unify the
implementation for table view and collection view, the `UserInterface`
now demands a different method `performUpdates( _ updateClosure: () ->
Void, completion: (() -> Void)?)`.

For table views the update closure is invoked in-between
`beginUpdates()` and `endUpdates`. For collection view the
`updateClosure` is called inside `performBatchUpdates`.

`beginUpdates` and `endUpdates` have been removed from `UserInterface`.